Reintroduced some bytecode (.class and .jar) files that are required by tcljava for building.
Those files had been removed in the expectation that they would be rebuilt by
the tcljava build script (Makefile), but removing them breaks the build.

Performing tcljava's 'make clean' doesn't remove these files.

It would be more satisfactory to be able to clean and rebuild these files
automatically, if possible. For the moment, they will be provided as is.

for easier investigation and future improvements, the list of these
files will be provided here:

  tcljava/demos/gluepkg/A_Jar_File.jar
  tcljava/src/Test.class
  tcljava/src/empty/empty.jar
  tcljava/src/tclblend/tcl/lang/CObject.class
  tcljava/src/tclblend/tcl/lang/IdleHandler.class
  tcljava/src/tclblend/tcl/lang/Interp.class
  tcljava/src/tclblend/tcl/lang/Notifier.class
  tcljava/src/tclblend/tcl/lang/TclList.class
  tcljava/src/tclblend/tcl/lang/TclObject.class
  tcljava/src/tclblend/tcl/lang/TimerHandler.class
  tcljava/src/tclblend/tcl/lang/Util.class
  tcljava/tests/tl.jar