tcljava/Makefile.in: bug fix: remove an @echo that I introduced for debugging in May 3, 2019
Oops...

(commit d2f63093a4ead1e5a239755696a42578c72df56d) when upgrading tcljava for JDK 10+
and that I forgot to remove after implementing javac -h.

(javah was superseded by javac -h since JDK 8, deprecated in JDK 9 and removed in JDK 10)

Not sure why it never failed on my systems, preventing me from noticing this earlier,
but it certainly broke the build in some older environments
(the effect of @echo was not only to display the javah call, but also to disable it).